New Portage 2.0.10 release with amazing new repoman -- and new x86 rescue portage

+++ sys-apps/portage/portage-2.0.10.ebuild
@@ -1,12 +1,15 @@
 # Copyright 1999-2002 Gentoo Technologies, Inc. Distributed under the terms
 # of the GNU General Public License, v2 or later 
-# $Header: /var/cvsroot/gentoo-x86/sys-apps/portage/Attic/portage-2.0.5.ebuild,v 1.1 2002/06/24 17:11:09 drobbins Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-apps/portage/Attic/portage-2.0.10.ebuild,v 1.1 2002/07/06 05:35:41 drobbins Exp $
  
 S=${WORKDIR}/${P}
 SLOT="0"
 DESCRIPTION="Portage ports system"
 SRC_URI=""
 HOMEPAGE="http://www.gentoo.org"
+KEYWORDS="*"
+LICENSE="GPL-2"
+
 #We need this if/then/else clause for compatibility with stuff that doesn't know !build?
 if [ "`use build`" ]
 then
@@ -149,15 +152,25 @@ pkg_postinst() {
 			cd ..
 		done
 		echo ">>> Database upgrade complete."
+		#remove old virtual directory to prevent virtual deps from getting messed-up
+		[ -d ${ROOT}var/db/pkg/virtual ] && rm -rf ${ROOT}var/db/pkg/virtual
 	fi
 
 	#fix cache (could contain staleness)
-	if [ -d ${ROOT}var/cache/edb/dep ]
+	if [ ! -d ${ROOT}var/cache/edb/dep/sys-apps ]
 	then
-		#avoid using "*" below as it can overwhelm rm
-		rm -rf ${ROOT}var/cache/edb/dep
-	fi	
-	install -d ${ROOT}var/cache/edb/dep
+		if [ -d ${ROOT}var/cache/edb/dep ]
+		then
+			#avoid using "*" below as it can overwhelm rm
+			rm -rf ${ROOT}var/cache/edb/dep
+		fi	
+		#ok, set setgid wheel on the cache directory so that "wheel" users can cache stuff too.
+		install -m2775 -o root -g wheel -d ${ROOT}var/cache/edb/dep
+	
+	else
+		chown -R root.wheel ${ROOT}var/cache/edb/dep/*
+		chmod g+sw ${ROOT}var/cache/edb/dep/*
+	fi
 	rm -f ${ROOT}usr/lib/python2.2/site-packages/portage.py[co]
 	# we gotta re-compile these modules and deal with systems with clock skew (stale compiled files)
 	python -c "import py_compile; py_compile.compile('${ROOT}usr/lib/python2.2/site-packages/portage.py')" || die
